+++ lib/commands/lock.js
@@ -112,15 +112,9 @@ export function validateUnlockCapabilities(caps) {
 }
 
 /**
- * @typedef {Object} FastUnlockOptions
- * @property {string} credential
- * @property {string} credentialType
- */
-
-/**
  *
  * @param {import('appium-adb').ADB} adb
- * @param {FastUnlockOptions} opts
+ * @param {import('./types').FastUnlockOptions} opts
  */
 async function fastUnlock(adb, opts) {
   const {credential, credentialType} = opts;
