+++ lib/helpers/device.js
@@ -1,54 +1,11 @@
-import ADB from 'appium-adb';
 import semver from 'semver';
 import _ from 'lodash';
-import { util } from '@appium/support';
-import logger from '../logger';
 import { resetMockLocation, setMockLocationApp } from '../commands/geolocation';
 import { pushSettingsApp } from '../commands/settings-app';
 import { SETTINGS_HELPER_ID } from 'io.appium.settings';
 import { ensureDeviceLocale } from '../commands/resources';
 import { hideKeyboardCompletely, initUnicodeKeyboard } from '../commands/keyboard';
-
-/**
- * @param {import('../driver').AndroidDriverOpts?} [opts=null]
- * @returns {Promise<ADB>}
- */
-async function createBaseADB(opts = null) {
-  // filter out any unwanted options sent in
-  // this list should be updated as ADB takes more arguments
-  const {
-    adbPort,
-    suppressKillServer,
-    remoteAdbHost,
-    clearDeviceLogsOnStart,
-    adbExecTimeout,
-    useKeystore,
-    keystorePath,
-    keystorePassword,
-    keyAlias,
-    keyPassword,
-    remoteAppsCacheLimit,
-    buildToolsVersion,
-    allowOfflineDevices,
-    allowDelayAdb,
-  } = opts ?? {};
-  return await ADB.createADB({
-    adbPort,
-    suppressKillServer,
-    remoteAdbHost,
-    clearDeviceLogsOnStart,
-    adbExecTimeout,
-    useKeystore,
-    keystorePath,
-    keystorePassword,
-    keyAlias,
-    keyPassword,
-    remoteAppsCacheLimit,
-    buildToolsVersion,
-    allowOfflineDevices,
-    allowDelayAdb,
-  });
-}
+import { createBaseADB, prepareEmulator, validatePackageActivityNames } from './utils';
 
 /**
  * @this {import('../driver').AndroidDriver}
@@ -59,8 +16,8 @@ export async function getDeviceInfoFromCaps() {
   // on instantiating on earlier (at this point, we have no udid)
   // we can only use this ADB object for commands that would not be confused
   // if multiple devices are connected
-  /** @type {ADB} */
-  const adb = await createBaseADB.bind(this)();
+  /** @type {import('appium-adb').ADB} */
+  const adb = this.adb ?? await createBaseADB(this.opts);
   let udid = this.opts.udid;
   let emPort;
 
@@ -174,7 +131,6 @@ export async function createADB(opts = null) {
   if (emPort) {
     adb.setEmulatorPort(emPort);
   }
-
   return adb;
 }
 
@@ -190,7 +146,7 @@ export async function getLaunchInfo() {
   let {appPackage, appActivity, appWaitPackage, appWaitActivity} = this.opts;
   const {app} = this.opts;
 
-  validatePackageActivityNames(this.opts);
+  validatePackageActivityNames.bind(this)();
 
   if (appPackage && appActivity) {
     return;
@@ -235,7 +191,7 @@ export async function initDevice() {
   } = this.opts;
 
   if (skipDeviceInitialization) {
-    logger.info(`'skipDeviceInitialization' is set. Skipping device initialization.`);
+    this.log.info(`'skipDeviceInitialization' is set. Skipping device initialization.`);
   } else {
     if (this.isEmulator()) {
       // Check if the device wake up only for an emulator.
@@ -272,7 +228,7 @@ export async function initDevice() {
   }
 
   if (skipLogcatCapture) {
-    logger.info(`'skipLogcatCapture' is set. Skipping starting logcat capture.`);
+    this.log.info(`'skipLogcatCapture' is set. Skipping starting logcat capture.`);
   } else {
     await this.adb.startLogcat({
       format: logcatFormat,
@@ -287,127 +243,10 @@ export async function initDevice() {
   }
 
   if (unicodeKeyboard) {
-    logger.warn(
+    this.log.warn(
       `The 'unicodeKeyboard' capability has been deprecated and will be removed. ` +
       `Set the 'hideKeyboard' capability to 'true' in order to make the on-screen keyboard invisible.`
     );
     return await initUnicodeKeyboard.bind(this)();
   }
 }
-
-// ***
-// #region Private Helpers
-// ***
-
-/**
- * @param {import('../driver').AndroidDriverOpts} opts
- * @returns {void}
- */
-function validatePackageActivityNames(opts) {
-  for (const key of ['appPackage', 'appActivity', 'appWaitPackage', 'appWaitActivity']) {
-    const name = opts[key];
-    if (!name) {
-      continue;
-    }
-
-    const match = /([^\w.*,])+/.exec(String(name));
-    if (!match) {
-      continue;
-    }
-
-    logger.warn(
-      `Capability '${key}' is expected to only include latin letters, digits, underscore, dot, comma and asterisk characters.`
-    );
-    logger.warn(
-      `Current value '${name}' has non-matching character at index ${match.index}: '${String(
-        name
-      ).substring(0, match.index + 1)}'`
-    );
-  }
-}
-
-/**
- * @this {import('../driver').AndroidDriver}
- * @param {string} networkSpeed
- * @returns {string}
- */
-function ensureNetworkSpeed(networkSpeed) {
-  if (networkSpeed.toUpperCase() in this.adb.NETWORK_SPEED) {
-    return networkSpeed;
-  }
-  logger.warn(
-    `Wrong network speed param '${networkSpeed}', using default: ${this.adb.NETWORK_SPEED.FULL}. ` +
-      `Supported values: ${_.values(this.adb.NETWORK_SPEED)}`
-  );
-  return this.adb.NETWORK_SPEED.FULL;
-}
-
-/**
- * @this {import('../driver').AndroidDriver}
- * @returns {string[]}
- */
-function prepareAvdArgs() {
-  const {networkSpeed, isHeadless, avdArgs} = this.opts;
-  const result = [];
-  if (avdArgs) {
-    if (_.isArray(avdArgs)) {
-      result.push(...avdArgs);
-    } else {
-      result.push(...(util.shellParse(`${avdArgs}`)));
-    }
-  }
-  if (networkSpeed) {
-    result.push('-netspeed', ensureNetworkSpeed.bind(this)(networkSpeed));
-  }
-  if (isHeadless) {
-    result.push('-no-window');
-  }
-  return result;
-}
-
-/**
- * @this {import('../driver').AndroidDriver}
- * @returns {Promise<void>}
- */
-async function prepareEmulator() {
-  const {
-    avd,
-    avdEnv: env,
-    language,
-    locale: country,
-    avdLaunchTimeout: launchTimeout,
-    avdReadyTimeout: readyTimeout,
-  } = this.opts;
-  if (!avd) {
-    throw new Error('Cannot launch AVD without AVD name');
-  }
-
-  const avdName = avd.replace('@', '');
-  let isEmulatorRunning = true;
-  try {
-    await this.adb.getRunningAVDWithRetry(avdName, 5000);
-  } catch (e) {
-    this.log.debug(`Emulator '${avdName}' is not running: ${e.message}`);
-    isEmulatorRunning = false;
-  }
-  const args = prepareAvdArgs.bind(this)();
-  if (isEmulatorRunning) {
-    if (args.includes('-wipe-data')) {
-      this.log.debug(`Killing '${avdName}' because it needs to be wiped at start.`);
-      await this.adb.killEmulator(avdName);
-    } else {
-      this.log.debug('Not launching AVD because it is already running.');
-      return;
-    }
-  }
-  await this.adb.launchAVD(avd, {
-    args,
-    env,
-    language,
-    country,
-    launchTimeout,
-    readyTimeout,
-  });
-}
-
-// #endregion
