+++ lib/index.ts
@@ -6,12 +6,6 @@ export {ANDROID_DRIVER_CONSTRAINTS as commonCapConstraints} from './constraints'
 export * from './driver';
 export * as doctor from './doctor/checks';
 export type * from './helpers/types';
-export {
-  CHROMIUM_WIN,
-  NATIVE_WIN,
-  WEBVIEW_BASE,
-  WEBVIEW_WIN,
-  default as webviewHelpers,
-} from './commands/context/helpers';
+export * as androidHelpers from './helpers/device';
 
 export default AndroidDriver;
